Add configurable Boot Metric Status wait (#151)
* Update klog_main.cpp

Added a system boot-up check function and call it in the main. It will query the windows api for boot metrics every 10 seconds and wait for it to say it is fully booted before executing the rest of the program code. I've very lightly tested this but it is not code breaking at all and appears to work in my virtual machine environments on boot ups.

* Add configurable boot wait defines so one can configure whether or not they want the boot wait feature enabled depending on their use case.

---------

Co-authored-by: mydarkthawts <mydarkthawts@mydarkthawts.com>